third_party/blink/web_tests/fullscreen/full-screen-exit-when-popup.html

<body>
<script src="full-screen-test.js"></script>
<span></span>
<script>
    if (window.testRunner) {
        testRunner.setPopupBlockingEnabled(false);
    }
    var callback;
    var fullscreenChanged = function(event)
    {
        if (callback)
            callback(event)
    };
    waitForEvent(document, 'webkitfullscreenchange', fullscreenChanged);

    var spanEnteredFullScreen = function() {
        testExpected("document.webkitIsFullScreen", true);
        testExpected("document.webkitCurrentFullScreenElement", span);
        callback = cancelledFullScreen;
        testExpected("window.open('about:blank')", null, "!=");
    };

    var cancelledFullScreen = function() {
        testExpected("document.webkitIsFullScreen", false);
        testExpected("document.webkitCurrentFullScreenElement", undefined);
        endTest();
    };

    var span = document.getElementsByTagName('span')[0];

    callback = spanEnteredFullScreen;
    runWithKeyDown(function(){span.webkitRequestFullScreen()});
</script>
