third_party/blink/web_tests/fullscreen/full-screen-exit-when-popup-expected.txt

EVENT(webkitfullscreenchange)
EXPECTED (document.webkitIsFullScreen == 'true') OK
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLSpanElement]') OK
EXPECTED (window.open('about:blank') != 'null') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitIsFullScreen == 'false') OK
EXPECTED (document.webkitCurrentFullScreenElement == 'undefined') OK
END OF TEST
